+++ src/index.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
+import Root from './Root'
 import App from 'components/App'
 import './index.scss'
 
@@ -9,4 +10,9 @@ import { config } from 'config'
 
 initApi(config.API_KEY)
 
-ReactDOM.render(<App />, document.getElementById('root'))
+ReactDOM.render(
+  <Root>
+    <App />
+  </Root>,
+  document.getElementById('root')
+)